model로 이동, 1이상 indent 로직 수정, Race가 Car의 속성을 가지도록 수정
