--- src/main/java/step3/Race.java
@@ -1,28 +1,0 @@
-package step3;
-
-import java.util.Random;
-
-public class Race {
-
-    private static final Random random = new Random();
-    private static final int MAX_RANDOM_VALUE = 10;
-
-    public static void playing(int cars, int times) {
-        String[] board = new RaceBoard(cars).generate();
-
-        for (int time = 1; time <= times; time++) {
-            for (int car = 1; car <= cars; car++) {
-                if (Car.isMovable(getRandomValue())) {
-                    board[car - 1] = new Car(board[car - 1]).go();
-                }
-            }
-
-            new RaceResult(board).show();
-        }
-    }
-
-    public static int getRandomValue() {
-        return random.nextInt(MAX_RANDOM_VALUE);
-    }
-}
-
